Fix getgrouplist for Cygwin
The cygwin implementation of getgrouplist will return the number of groups of
a user even if the group array is too small (or null); this differs from
the behavior of the linux call, where the return value in that case is -1.

Use a retval > 0 check for cygwin (and keep the < 0 relation for other cases).