Adapt test 01 for cygwin environment
Test 01-getgrouplist-for-root.t checked whether user "root" belongs to group
0 (root as well). This obviously is not the case in Cygwin environments, where
user root does not exist. However, there's an "Administrator" with a uid of 500.

This commit checks user root if uid 0 exists, and tries the username of uid 500
otherwise.

This should still be valid for unix/linux, but also work with Cygwin.

A new test 02 was added that checks that the code does not crash for a
non-existing user ("barfuser").